+++ technologic/lib/technologic/logger.rb
@@ -15,6 +15,7 @@ module Technologic
         # `#to_log_string` is idiomatic to Technologic and is the most explicit way to specify how an object is logged.
         return value.to_log_string if value.respond_to?(:to_log_string)
 
+        return value.to_s if value.is_a?(Enumerator)
         return value if value.is_a?(Numeric)
         return value.id if value.respond_to?(:id)
         return value.map { |mappable_value| format_value_for_log(mappable_value) } if value.respond_to?(:map)